requirements - Copy.txt
streamlit>=1.30.0
pandas>=2.0.0
altair>=5.0.0
mysql-connector-python>=8.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
kafka-python>=2.0.0
